requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
phonenumbers==8.13.50
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pytz==2024.2
sqlparse==0.5.2
